+++ src/Contract/JsonSchema/Describer/SchemaAttributeDescriber.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\AI\Platform\Contract\JsonSchema\Describer;
 
-use Symfony\AI\Platform\Contract\JsonSchema\Attribute\With;
+use Symfony\AI\Platform\Contract\JsonSchema\Attribute\Schema;
 use Symfony\AI\Platform\Contract\JsonSchema\Factory;
 use Symfony\AI\Platform\Contract\JsonSchema\Subject\PropertySubject;
 use Symfony\AI\Platform\Exception\IOException;
@@ -19,22 +19,22 @@ use Symfony\AI\Platform\Exception\IOException;
 /**
  * @phpstan-import-type JsonSchema from Factory
  */
-final class WithAttributeDescriber implements PropertyDescriberInterface
+final class SchemaAttributeDescriber implements PropertyDescriberInterface
 {
     public function describeProperty(PropertySubject $subject, ?array &$schema): void
     {
-        foreach ($subject->getAttributes(With::class) as $attribute) {
+        foreach ($subject->getAttributes(Schema::class) as $attribute) {
             if ($attribute->ref) {
                 try {
-                    $with = json_decode(file_get_contents($attribute->ref), true, flags: \JSON_THROW_ON_ERROR);
+                    $attributeSchema = json_decode(file_get_contents($attribute->ref), true, flags: \JSON_THROW_ON_ERROR);
                 } catch (\JsonException $e) {
                     throw new IOException(\sprintf('Failed to load the schema from "%s"', $attribute->ref), 0, $e);
                 }
             } else {
-                $with = array_filter((array) $attribute, static fn ($value) => null !== $value);
+                $attributeSchema = array_filter((array) $attribute, static fn ($value) => null !== $value);
             }
 
-            $schema = array_replace_recursive($schema ?? [], $with);
+            $schema = array_replace_recursive($schema ?? [], $attributeSchema);
         }
     }
 }